+++ unity/FirstPersonController/Assets/DSS/FirstPersonController/CHANGELOG.md
@@ -2,4 +2,9 @@
 
 ## 1.0.0
 
-Initial release.
+- Initial release.
+
+## 1.0.1
+
+- Added a `.paused` variable to the `FirstPersonController` script.
+- Added a `CursorController` script.